+++ src/main/java/fpt/swp391/carrentalsystem/controller/owner/OwnerCarImageController.java
@@ -21,7 +21,7 @@ public class OwnerCarImageController {
     private final CarImageService carImageService;
 
     @GetMapping("/{carId}/images")
-    public String manageImages(@PathVariable Long carId, Model model) {
+    public String manageImages(@PathVariable Integer carId, Model model) {
         User owner = getCurrentUser();
         model.addAttribute("carId", carId);
         model.addAttribute("images", carImageService.getImagesByCar(carId, owner));
@@ -29,7 +29,7 @@ public class OwnerCarImageController {
     }
 
     @PostMapping("/{carId}/images/upload")
-    public String uploadImages(@PathVariable Long carId,
+    public String uploadImages(@PathVariable Integer carId,
                                @RequestParam("files") List<MultipartFile> files) throws Exception {
         if (files != null && !files.isEmpty()) {
             carImageService.uploadImages(carId, getCurrentUser(), files);
@@ -38,13 +38,13 @@ public class OwnerCarImageController {
     }
 
     @PostMapping("/images/{imageId}/set-main")
-    public String setMain(@PathVariable Long imageId, @RequestParam Long carId) {
+    public String setMain(@PathVariable Long imageId, @RequestParam Integer carId) {
         carImageService.setMainImage(imageId, getCurrentUser());
         return "redirect:/owner/cars/" + carId + "/images";
     }
 
     @PostMapping("/images/{imageId}/delete")
-    public String delete(@PathVariable Long imageId, @RequestParam Long carId) {
+    public String delete(@PathVariable Long imageId, @RequestParam Integer carId) {
         carImageService.deleteImage(imageId, getCurrentUser());
         return "redirect:/owner/cars/" + carId + "/images";
     }
