+++ src/main/java/fpt/swp391/carrentalsystem/service/impl/CarServiceByThanhQCImpl.java
@@ -69,7 +69,7 @@ public class CarServiceByThanhQCImpl implements CarServiceByThanhQC {
     @Override public List<String> getAllCarTypes() { return carRepository.findDistinctCarTypes(); }
     @Override public List<String> getAllFuelTypes() { return carRepository.findDistinctFuelTypes(); }
     @Override public List<Integer> getAllSeats() { return carRepository.findDistinctSeats(); }
-    @Override public Car getCarById(Long id) { return carRepository.findByIdWithImages(id).orElse(null); }
+    @Override public Car getCarById(Integer id) { return carRepository.findByIdWithImages(id).orElse(null); }
 
     @Override public List<CarListItemResponse> getCarsByOwner(Long ownerId) {
         return carRepository.findByOwner_Id(ownerId).stream().map(carMapper::toListItemResponse).collect(Collectors.toList());
@@ -85,14 +85,14 @@ public class CarServiceByThanhQCImpl implements CarServiceByThanhQC {
     }
 
     @Override @Transactional
-    public void approveCar(Long id) {
+    public void approveCar(Integer id) {
         Car car = carRepository.findById(id).orElseThrow(() -> new RuntimeException("Car not found"));
         car.setStatus(CarStatus.AVAILABLE);
         carRepository.save(car);
     }
 
     @Override @Transactional
-    public void rejectCar(Long id) {
+    public void rejectCar(Integer id) {
         Car car = carRepository.findById(id).orElseThrow(() -> new RuntimeException("Car not found"));
         car.setStatus(CarStatus.REJECTED);
         carRepository.save(car);
@@ -100,7 +100,7 @@ public class CarServiceByThanhQCImpl implements CarServiceByThanhQC {
 
     @Override
     @Transactional
-    public void updateCarStatus(Long id, CarStatus status) {
+    public void updateCarStatus(Integer id, CarStatus status) {
         Car car = carRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Không tìm thấy xe"));
         car.setStatus(status);
